add model builder and refactor apis
