+++ lib/sls-api-file-delete.js
@@ -1,31 +1,11 @@
 'use strict';
 
 const { API } = require('@janiscommerce/api');
-const Model = require('@janiscommerce/model');
+const ModelBuilder = require('./model-builder');
 const SlsApiFileDeleteError = require('./sls-api-file-delete-error');
 // const S3 = require('@janiscommerce/s3');
 
 class SlsApiFileDelete extends API {
-
-	/**
-	 * Build a model class with databaseKey and table passed in params
-	 * @param {string} table
-	 * @param {string} dataBaseKey
-	 */
-	static buildModel(table, dataBaseKey) {
-		class FileModel extends Model {
-			get databaseKey() {
-				return dataBaseKey;
-			}
-
-			static get table() {
-				return table;
-			}
-		}
-
-		return FileModel;
-	}
-
 
 	/**
 	 * Database table where to save the record
@@ -67,7 +47,7 @@ class SlsApiFileDelete extends API {
 
 		const [entityId, fileId] = this.pathParameters;
 
-		const model = this.constructor.buildModel(this.table, this.databaseKey);
+		const model = ModelBuilder.build(this.table, this.databaseKey, this.entityIdField);
 
 		this.modelInstance = this.session.getSessionInstance(model);
