+++ lib/sls-api-file-relation.js
@@ -2,8 +2,8 @@
 
 const { struct } = require('superstruct');
 const { API } = require('@janiscommerce/api');
-const Model = require('@janiscommerce/model');
 const S3 = require('@janiscommerce/s3');
+const ModelBuilder = require('./model-builder');
 const SlsApiFileRelationError = require('./sls-api-file-relation-error');
 
 class SlsApiFileRelation extends API {
@@ -34,27 +34,6 @@ class SlsApiFileRelation extends API {
 				return 'other';
 		}
 	}
-
-
-	/**
-	 * Build a model class with databaseKey and table passed in params
-	 * @param {string} table
-	 * @param {string} dataBaseKey
-	 */
-	static buildModel(table, dataBaseKey) {
-		class FileModel extends Model {
-			get databaseKey() {
-				return dataBaseKey;
-			}
-
-			static get table() {
-				return table;
-			}
-		}
-
-		return FileModel;
-	}
-
 
 	/**
 	 * Database table where to save the record
@@ -128,7 +107,7 @@ class SlsApiFileRelation extends API {
 
 		const { fileName: name, fileSource: path, ...additionalData } = this.data;
 
-		const model = this.constructor.buildModel(this.table, this.databaseKey);
+		const model = ModelBuilder.build(this.table, this.databaseKey, this.entityIdField);
 
 		const modelInstance = this.session.getSessionInstance(model);
